Fix for missing clusters on printout
Fix for missing clusters on printout when Print Task is used.
The _opacity property is added to give the user the opportunity to make the clusters transparent. Without setting the opacity on the Cluster layer itself, the clusters are not transparent. I personally, changed the cluster symbol colors and made them not transparent to have the same result on the printout. But this is a personal preference and I did not include it here. 
Also, if a ClassBreaksRenderer is set after the ClusterFeatureLayer is instantiated, the Print Task throws an error. So, in order for the Print Task to work, the provided ClassBreaksRenderer must be used.